requirements.txt
appdirs==1.4.0
blinker==1.4
boto==2.46.1
bz2file==0.98
click==6.7
coverage==4.3.4
decorator==4.0.11
enum-compat==0.0.2
eventlet==0.20.1
flake8==3.3.0
Flask==0.12
Flask-DebugToolbar==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SocketIO==2.8.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
futures==3.0.5
gensim==1.0.1
greenlet==0.4.12
IndicoIo==1.0.8
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
mccabe==0.6.1
mock==1.3.0
numpy==1.12.1
olefile==0.44
packaging==16.8
passlib==1.7.1
pbr==1.10.0
Pillow==4.1.0 --install-option="build_ext" --install-option="--disable-jpeg" --install-option="--disable-zlib"
psycopg2==2.6.2
py==1.4.32
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.1.10
pytest==3.0.6
pytest-cov==2.4.0
pytest-flask==0.10.0
python-engineio==1.1.2
python-socketio==1.6.3
requests==2.13.0
scipy==0.19.0
six==1.10.0
smart-open==1.5.1
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.15
WTForms==2.1
